saudinews/requirements.txt
Flask==3.0.3
feedparser==6.0.11
newspaper3k==0.2.8
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0
